requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
PyGithub==2.1.1
python-multipart==0.0.6
pydantic==2.5.0
